backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-environ>=0.11.2",
    "django>=5.1.4",
    "django-ninja>=1.3.0",
    "psycopg>=3.2.3",
    "django-ninja-jwt>=5.3.5",
]
